requirements.txt
openai>=1.30.0
markdown>=3.6
requests>=2.31.0
python-dotenv>=1.0.1
